+++ src/lib/components/upload/UploadSection.tsx
@@ -132,7 +132,8 @@ export const UploadSection = ({
         name="codeDesc"
         control={control}
         label="Code Description (Optional)"
-        helperText="Short description of your code. You can add or change this later."
+        placeholder="No Description"
+        helperText="Define what your code works on in one sentence which visible to you only. You can add this later."
         rules={{
           maxLength: MAX_CODE_DESCRIPTION_LENGTH,
         }}