+++ src/lib/pages/instantiate/instantiate.tsx
@@ -12,6 +12,7 @@ import {
   useNativeTokensInfo,
   useSimulateFee,
   useInstantiateTx,
+  useCelatoneApp,
 } from "lib/app-provider";
 import { CodeSelectSection } from "lib/components/CodeSelectSection";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
@@ -48,6 +49,9 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   const router = useRouter();
   const msgQuery = (router.query.msg as string) ?? "";
   const codeIdQuery = (router.query["code-id"] as string) ?? "";
+  const {
+    appContractAddress: { example: exampleContractAddress },
+  } = useCelatoneApp();
   const { address = "" } = useWallet();
   const endpoint = useLCDEndpoint();
   const postInstantiateTx = useInstantiateTx();
@@ -275,6 +279,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             name="label"
             control={control}
             error={formErrors.label?.message}
+            placeholder="ex. Token Factory"
             label="Label"
             helperText="Label will help remind you or other contract viewer to understand what this contract do and how it works"
             variant="floating"
@@ -285,6 +290,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             name="adminAddress"
             control={control}
             label="Admin Address (optional)"
+            placeholder={`ex. ${exampleContractAddress}`}
             helperText="This address will be the admin for the deployed smart contract."
             variant="floating"
             error={validateAdmin(watchAdminAddress)}